Refactored integration test boilerplate to base class
